requirements.txt
fastapi
uvicorn
requests
PyYAML
ratelimit